// graphql-api/lib/interfaces/merge-report.interface.ts
export interface IBorrowerTelephone extends IPartitionElements {
  PhoneNumber: IPhoneNumber;
  PhoneType: ICodeRef;
  Source: ISource;
}
export interface IPhoneNumber {
  AreaCode: string | null;
  Number: string | null;
  Extension: string | null;
}